Rename com.ibm.ws.security.auth.data_fat to io.openliberty
