+++ dev/io.openliberty.security.auth.data_fat/.settings/bndtools.core.prefs
