+++ dev/io.openliberty.security.auth.data_fat/fat/src/io/openliberty/security/auth/data/fat/AuthDataFatUtils.java
@@ -1,0 +1,55 @@
+/*******************************************************************************
+ * Copyright (c) 2020, 2026 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+
+package io.openliberty.security.auth.data.fat;
+
+import java.io.File;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+import com.ibm.websphere.simplicity.config.ServerConfiguration;
+
+import componenttest.rules.repeater.JakartaEEAction;
+import componenttest.topology.impl.LibertyServer;
+
+/**
+ * Some utility functions for FATs.
+ */
+public class AuthDataFatUtils {
+
+    /**
+     * JakartaEE9 transform a list of applications. The applications are the simple app names and they must exist at '<server>/apps/<appname>'.
+     *
+     * @param myServer The server to transform the applications on.
+     * @param apps     The simple names of the applications to transform.
+     */
+    public static void transformApps(LibertyServer myServer, String... apps) {
+        if (JakartaEEAction.isEE9OrLaterActive()) {
+            for (String app : apps) {
+                Path someArchive = Paths.get(myServer.getServerRoot() + File.separatorChar + "apps" + File.separatorChar + app);
+                JakartaEEAction.transformApp(someArchive);
+            }
+        }
+    }
+
+    /**
+     * Swap out the passwordUtilities-1.0 feature with the passwordUtilities-1.1 feature.
+     *
+     * @param server The server running with passwordUtilities-1.0.
+     * @throws Exception if there was an issue swapping the features.
+     */
+    public static void runWithPasswordUtilities11(LibertyServer server) throws Exception {
+        ServerConfiguration config = server.getServerConfiguration().clone();
+        config.getFeatureManager().getFeatures().remove("passwordUtilities-1.0");
+        config.getFeatureManager().getFeatures().remove("passwordutilities-1.0"); // JakartaEE9Action lower-cases feature names
+        config.getFeatureManager().getFeatures().add("passwordUtilities-1.1");
+        server.updateServerConfiguration(config);
+    }
+}
