+++ dev/io.openliberty.security.auth.data_fat/publish/servers/io.openliberty.security.auth.data.fat.dpm/server.xml
