+++ dev/io.openliberty.security.auth.data_fat/test-applications/DefaultPrincipalMappingApp.war/resources/META-INF/permissions.xml
