+++ dev/io.openliberty.security.auth.data_fat/test-applications/DefaultPrincipalMappingApp.war/resources/WEB-INF/web.xml
