+++ dev/io.openliberty.security.auth.data_fat/test-applications/DefaultPrincipalMappingApp.war/src/web/DefaultPrincipalMappingServlet.java
